test: reuse test_backward code to implement triton_tester
